# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: LegalDesk — Legal Office Management Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /documents:
    get:
      operationId: listDocuments
      tags: [documents]
      summary: List documents
      parameters:
        - name: caseId
          in: query
          required: false
          schema:
            type: integer
        - name: clientId
          in: query
          required: false
          schema:
            type: integer
        - name: docType
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"
    post:
      operationId: createDocument
      tags: [documents]
      summary: Upload/create a document
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDocumentBody"
      responses:
        "201":
          description: Created document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
  /documents/{id}:
    get:
      operationId: getDocument
      tags: [documents]
      summary: Get a document by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        "404":
          description: Not found
    delete:
      operationId: deleteDocument
      tags: [documents]
      summary: Delete a document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # TASKS
...
  /hearings:
    get:
      operationId: listHearings
      tags: [hearings]
      summary: List hearings/sessions
      parameters:
        - name: caseId
          in: query
          required: false
          schema:
            type: integer
        - name: assignedLawyer
          in: query
          required: false
          schema:
            type: integer
        - name: status
          in: query
          required: false
          schema:
            type: string
        - name: upcoming
          in: query
          required: false
          schema:
            type: boolean
      responses:
        "200":
          description: List of hearings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Hearing"
    post:
      operationId: createHearing
      tags: [hearings]
      summary: Create a new hearing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateHearingBody"
      responses:
        "201":
          description: Created hearing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Hearing"
  /hearings/{id}:
    get:
      operationId: getHearing
      tags: [hearings]
      summary: Get a hearing by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Hearing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Hearing"
        "404":
          description: Not found
    patch:
      operationId: updateHearing
      tags: [hearings]
      summary: Update a hearing
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateHearingBody"
      responses:
        "200":
          description: Updated hearing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Hearing"
    delete:
      operationId: deleteHearing
      tags: [hearings]
      summary: Delete a hearing
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # CONSULTATIONS
...
  /payments:
    get:
      operationId: listPayments
      tags: [payments]
      summary: List payments
      parameters:
        - name: clientId
          in: query
          required: false
          schema:
            type: integer
        - name: caseId
          in: query
          required: false
          schema:
            type: integer
        - name: status
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of payments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Payment"
    post:
      operationId: createPayment
      tags: [payments]
      summary: Record a payment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePaymentBody"
      responses:
        "201":
          description: Created payment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
  /payments/{id}:
    get:
      operationId: getPayment
      tags: [payments]
      summary: Get a payment by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Payment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
        "404":
          description: Not found
    patch:
      operationId: updatePayment
      tags: [payments]
      summary: Update a payment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePaymentBody"
      responses:
        "200":
          description: Updated payment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"

  # POWERS OF ATTORNEY
...
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: List notifications for current user
      parameters:
        - name: unreadOnly
          in: query
          required: false
          schema:
            type: boolean
      responses:
        "200":
          description: List of notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark a notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated notification
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
...
components:
  schemas:
...
    Document:
      type: object
      properties:
        id:
          type: integer
        caseId:
          type: ["integer", "null"]
        clientId:
          type: ["integer", "null"]
        fileUrl:
          type: string
        fileName:
          type: string
        docType:
          type: string
          enum: [contract, ruling, memo, power_of_attorney, id_copy, other]
        isOriginal:
          type: boolean
        uploadedBy:
          type: ["integer", "null"]
        uploaderName:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - fileUrl
        - fileName
        - docType
        - isOriginal
        - createdAt
    CreateDocumentBody:
      type: object
      properties:
        caseId:
          type: ["integer", "null"]
        clientId:
          type: ["integer", "null"]
        fileUrl:
          type: string
        fileName:
          type: string
        docType:
          type: string
          enum: [contract, ruling, memo, power_of_attorney, id_copy, other]
        isOriginal:
          type: boolean
        notes:
          type: ["string", "null"]
      required:
        - fileUrl
        - fileName
        - docType

    # TASK
...
    Hearing:
      type: object
      properties:
        id:
          type: integer
        caseId:
          type: integer
        caseNumber:
          type: ["string", "null"]
        datetime:
          type: string
          format: date-time
        court:
          type: ["string", "null"]
        type:
          type: string
          enum: [session, consultation, deadline, other]
        assignedLawyer:
          type: ["integer", "null"]
        assignedLawyerName:
          type: ["string", "null"]
        status:
          type: string
          enum: [scheduled, completed, adjourned, cancelled]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - caseId
        - datetime
        - type
        - status
        - createdAt
    CreateHearingBody:
      type: object
      properties:
        caseId:
          type: integer
        datetime:
          type: string
          format: date-time
        court:
          type: ["string", "null"]
        type:
          type: string
          enum: [session, consultation, deadline, other]
        assignedLawyer:
          type: ["integer", "null"]
        status:
          type: string
          enum: [scheduled, completed, adjourned, cancelled]
        notes:
          type: ["string", "null"]
      required:
        - caseId
        - datetime
        - type
    UpdateHearingBody:
      type: object
      properties:
        datetime:
          type: string
          format: date-time
        court:
          type: ["string", "null"]
        type:
          type: string
          enum: [session, consultation, deadline, other]
        assignedLawyer:
          type: ["integer", "null"]
        status:
          type: string
          enum: [scheduled, completed, adjourned, cancelled]
        notes:
          type: ["string", "null"]

    # CONSULTATION
...
    Payment:
      type: object
      properties:
        id:
          type: integer
        clientId:
          type: integer
        clientName:
          type: ["string", "null"]
        caseId:
          type: ["integer", "null"]
        caseNumber:
          type: ["string", "null"]
        consultationId:
          type: ["integer", "null"]
        amount:
          type: number
        type:
          type: string
          enum: [case_fee, consultation_fee, retainer, expense, other]
        status:
          type: string
          enum: [pending, paid, partial, overdue]
        paidAt:
          type: ["string", "null"]
          format: date-time
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - clientId
        - amount
        - type
        - status
        - createdAt
    CreatePaymentBody:
      type: object
      properties:
        clientId:
          type: integer
        caseId:
          type: ["integer", "null"]
        consultationId:
          type: ["integer", "null"]
        amount:
          type: number
        type:
          type: string
          enum: [case_fee, consultation_fee, retainer, expense, other]
        status:
          type: string
          enum: [pending, paid, partial, overdue]
        paidAt:
          type: ["string", "null"]
          format: date-time
        notes:
          type: ["string", "null"]
      required:
        - clientId
        - amount
        - type
    UpdatePaymentBody:
      type: object
      properties:
        amount:
          type: number
        type:
          type: string
          enum: [case_fee, consultation_fee, retainer, expense, other]
        status:
          type: string
          enum: [pending, paid, partial, overdue]
        paidAt:
          type: ["string", "null"]
          format: date-time
        notes:
          type: ["string", "null"]

    # POWER OF ATTORNEY
...
    Notification:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        type:
          type: string
          enum: [hearing_reminder, task_overdue, poa_overdue, new_consultation, payment_due, document_uploaded, task_update, general]
        title:
          type: string
        body:
          type: string
        read:
          type: boolean
        refId:
          type: ["integer", "null"]
        refType:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - type
        - title
        - body
        - read
        - createdAt

    # DASHBOARD
...